triggers created for due date table

+++ database.py
@@ -26,6 +26,12 @@ def read_customer():
 
 
 def delete_customer_db(cust_id):
+    #check if cust_id exists
+    mycursor.execute("SELECT * FROM customer WHERE cust_id = %s",(cust_id,))
+    myresult = mycursor.fetchall()
+    if not myresult:
+        st.write("cust_id does not exist")
+        return
     sql = "DELETE FROM customer WHERE cust_id = %s"
     val = (cust_id,)
     mycursor.execute(sql, val)
@@ -63,16 +69,22 @@ def update_admin_db(admin_id,name,Customer_id):
     st.write("Admin updated successfully")
 
 def delete_admin_db(admin_id):
+    #check if admin_id exists
+    mycursor.execute("SELECT * FROM admin WHERE admin_id = %s",(admin_id,))
+    myresult = mycursor.fetchall()
+    if not myresult:
+        st.write("admin_id does not exist")
+        return
     sql = "DELETE FROM admin WHERE admin_id = %s"
     val = (admin_id,)
     mycursor.execute(sql, val)
     mydb.commit()
     st.write("Admin deleted successfully")
 
-def add_bill(cust_id,meter_no,units,cost_per_unit):
-    mycursor.execute("CREATE TABLE IF NOT EXISTS bill (bill_id INT AUTO_INCREMENT PRIMARY KEY, cust_id INT NOT NULL, meter_no INT NOT NULL, units INT NOT NULL, cost_per_unit INT NOT NULL, amount INT NOT NULL, FOREIGN KEY (cust_id) REFERENCES customer(cust_id))")
-    sql = "INSERT INTO bill (cust_id, meter_no, units, cost_per_unit, amount) VALUES (%s, %s, %s, %s, %s)"
-    val = (cust_id, meter_no, units, cost_per_unit, int(units)*int(cost_per_unit))
+def add_bill(cust_id,meter_no,units,cost_per_unit,due_date):
+    mycursor.execute("CREATE TABLE IF NOT EXISTS bill (bill_id INT AUTO_INCREMENT PRIMARY KEY, cust_id INT NOT NULL, meter_no INT NOT NULL, units INT NOT NULL, cost_per_unit INT NOT NULL, amount INT NOT NULL, due_date DATE NOT NULL, FOREIGN KEY (cust_id) REFERENCES customer(cust_id))")
+    sql = "INSERT INTO bill (cust_id, meter_no, units, cost_per_unit, amount, due_date) VALUES (%s, %s, %s, %s, %s, %s)"
+    val = (cust_id, meter_no, units, cost_per_unit, int(units)*int(cost_per_unit), due_date)
     mycursor.execute(sql, val)
     mydb.commit()
     st.write("Bill added successfully")
@@ -82,18 +94,24 @@ def read_bill():
     mycursor.execute("SELECT * FROM bill")
     myresult = mycursor.fetchall()
     st.write("Bill details")
-    df=pd.DataFrame(myresult,columns=['bill_id','cust_id','meter_no','units','cost_per_unit','amount'])
+    df=pd.DataFrame(myresult,columns=['bill_id','cust_id','meter_no','units','cost_per_unit','amount','due_date'])
     st.dataframe(df)
 
 
-def update_bill_db(bill_id,cust_id,meter_no,units,cost_per_unit):
-    sql = "UPDATE bill SET cust_id = %s, meter_no = %s, units = %s, cost_per_unit = %s, amount = %s WHERE bill_id = %s"
-    val = (cust_id, meter_no, units, cost_per_unit, int(units)*int(cost_per_unit), bill_id)
+def update_bill_db(bill_id,cust_id,meter_no,units,cost_per_unit, due_date):
+    sql = "UPDATE bill SET cust_id = %s, meter_no = %s, units = %s, cost_per_unit = %s, amount = %s, due_date=%s WHERE bill_id = %s"
+    val = (cust_id, meter_no, units, cost_per_unit, int(units)*int(cost_per_unit), due_date, bill_id)
     mycursor.execute(sql, val)
     mydb.commit()
     st.write("Bill updated successfully")
 
 def delete_bill_db(bill_id):
+    #check if bill_id exists
+    mycursor.execute("SELECT * FROM bill WHERE bill_id = %s",(bill_id,))
+    myresult = mycursor.fetchall()
+    if not myresult:
+        st.write("bill_id does not exist")
+        return
     sql = "DELETE FROM bill WHERE bill_id = %s"
     val = (bill_id,)
     mycursor.execute(sql, val)
